requirements.txt
openai == 0.27.4
langchain == 0.0.135
pandas == 1.1.3
mysql-connector-python == 8.0.33
protobuf==3.14.0
six==1.15.0
grpcio == 1.42.0
grpcio-tools == 1.42.0
 